+++ app/src/androidTest/java/woowacourse/movie/presentation/fixture/FakeContext.kt
@@ -1,0 +1,6 @@
+package woowacourse.movie.presentation.fixture
+
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
+
+val fakeContext: Context = ApplicationProvider.getApplicationContext()
