+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".view.movies.MoviesActivity"
+            android:name=".presentation.view.movies.MoviesActivity"
             android:exported="true"
             android:launchMode="singleTask">
             <intent-filter>
@@ -24,10 +24,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".view.reservation.result.ReservationResultActivity"
+            android:name=".presentation.view.reservation.result.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".view.reservation.ReservationActivity"
+            android:name=".presentation.view.reservation.reservation.ReservationActivity"
             android:exported="false" />
     </application>
