+++ app/src/main/java/woowacourse/movie/presentation/extension/DateTimeExtension.kt
@@ -1,0 +1,8 @@
+package woowacourse.movie.presentation.extension
+
+import java.time.format.DateTimeFormatter
+
+fun String.toDateTimeFormatter(): DateTimeFormatter? =
+    runCatching {
+        DateTimeFormatter.ofPattern(this)
+    }.getOrNull()
