+++ app/src/main/java/woowacourse/movie/presentation/view/movies/MoviesPresenter.kt
@@ -1,0 +1,14 @@
+package woowacourse.movie.presentation.view.movies
+
+import woowacourse.movie.presentation.fixture.dummyMovie
+
+class MoviesPresenter(
+    private val view: MoviesContract.View,
+) : MoviesContract.Presenter {
+    override fun fetchData() {
+        val movies = listOf(dummyMovie)
+        view.setScreen(movies) { movie ->
+            view.navigateToReservationScreen(movie)
+        }
+    }
+}
