+++ app/src/main/java/woowacourse/movie/presentation/view/movies/OnMovieEventListener.kt
@@ -1,0 +1,7 @@
+package woowacourse.movie.presentation.view.movies
+
+import woowacourse.movie.domain.model.Movie
+
+interface OnMovieEventListener {
+    fun onClick(movie: Movie)
+}
