--- app/src/main/java/woowacourse/movie/view/extension/DateTimeExtension.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.view.extension
-
-import java.time.format.DateTimeFormatter
-
-fun String.toDateTimeFormatter(): DateTimeFormatter? =
-    runCatching {
-        DateTimeFormatter.ofPattern(this)
-    }.getOrNull()
