--- app/src/main/java/woowacourse/movie/view/movies/MoviesPresenter.kt
@@ -1,14 +1,0 @@
-package woowacourse.movie.view.movies
-
-import woowacourse.movie.view.fixture.dummyMovie
-
-class MoviesPresenter(
-    private val view: MoviesContract.View,
-) : MoviesContract.Presenter {
-    override fun fetchData() {
-        val movies = listOf(dummyMovie)
-        view.setScreen(movies) { movie ->
-            view.navigateToReservationScreen(movie)
-        }
-    }
-}
