--- app/src/main/java/woowacourse/movie/view/movies/OnMovieEventListener.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.view.movies
-
-import woowacourse.movie.domain.model.Movie
-
-interface OnMovieEventListener {
-    fun onClick(movie: Movie)
-}
