+++ app/src/main/res/layout/activity_movies.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.movies.MoviesActivity">
+    tools:context=".presentation.view.movies.MoviesActivity">
 
     <ListView
         android:id="@+id/lv_movie"
