+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.reservation.result.ReservationResultActivity">
+    tools:context=".presentation.view.reservation.result.ReservationResultActivity">
 
     <TextView
         android:id="@+id/tv_cancel_description"
